+++ lib/rubocop/cop/team.rb
@@ -74,7 +74,7 @@ module RuboCop
           @options[:stdin] = new_source
         else
           filename = buffer.name
-          File.open(filename, 'w') { |f| f.write(new_source) }
+          File.open(filename, 'wb') { |f| f.write(new_source) }
         end
         @updated_source_file = true
       end