+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug fixes
 
 [#4411](https://github.com/bbatsov/rubocop/issues/4411): Handle properly safe navigation in `Style/YodaCondition`. ([@bbatsov])
+Handle properly class variables and global variables in `Style/YodaCondition`. ([@bbatsov])
 
 ## 0.49.0 (2017-05-24)