+++ CHANGELOG.md
@@ -66,6 +66,9 @@
 * [#4394](https://github.com/bbatsov/rubocop/issues/4394): Prevent some cops from breaking on safe navigation operator. ([@drenmi][])
 * [#4260](https://github.com/bbatsov/rubocop/issues/4260): Prevent `Rails/SkipsModelValidations` from registering an offense for `FileUtils.touch`. ([@rrosenblum][])
 
+* [#4374](https://github.com/bbatsov/rubocop/issues/4374): Fix Regression of [#1566](https://github.com/bbatsov/rubocop/issues/1566) --auto-correct on windows changes the line-ending ([@dreamnite][])
+* [#4374](https://github.com/bbatsov/rubocop/issues/4374): Fix Regression of [#1566](https://github.com/bbatsov/rubocop/issues/1566) --auto-correct on windows changes the line-ending. ([@dreamnite][])
+ 
 ## 0.48.1 (2017-04-03)
 
 ### Changes
@@ -2800,3 +2803,4 @@
 [@gprado]: https://github.com/gprado
 [@yhirano55]: https://github.com/yhirano55
 [@hoshinotsuyoshi]: https://github.com/hoshinotsuyoshi
+[@dreamnite]: https://github.com/dreamnite